Cast enum to the an integer type before streaming out to QDataStream
We stream in as int or as qint32, so this is symmetric.

+++ share/qtcreator/qml/qmlpuppet/container/instancecontainer.cpp
@@ -99,8 +99,8 @@ QDataStream &operator<<(QDataStream &out, const InstanceContainer &container)
     out << container.minorNumber();
     out << container.componentPath();
     out << container.nodeSource();
-    out << container.nodeSourceType();
-    out << container.metaType();
+    out << qint32(container.nodeSourceType());
+    out << qint32(container.metaType());
 
     return out;
 }
